moved color measure methods
